rename(api-detail): mock 폴더 네이밍 수정
