+++ apps/monitor-web/src/mock/apiDetail.ts
