+++ apps/monitor-web/src/mock/index.ts
@@ -1,1 +1,1 @@
-export * from "./dashboard";
+export * from "./apiDetail";
